% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Test.R
\name{add}
\alias{add}
\title{This is a function for add}
\usage{
add(x, y)
}
\arguments{
\item{x}{is first parameter}

\item{y}{is second parameter}
}
\value{
x+y
}
\description{
This is a function for add
}
\author{
JijunYu
}
